Use objects instead of tilemap for ore
I thought it would be easier as a tilemap, but really, all that solved
was ensuring that the tiles were drawn in the correct spot, but I can do
that myself and not make it super complicated.

Additionally, I made it so we are always in a "multiplayer game" so we
can just use rpc calls instead of having to check if we're in
multiplayer or not to choose how to call the functions.