Add house interior and way to transition between scenes
In reality we just teleport the player between second level nodes on the
node tree, and have all nodes always loaded. This could be an issue in
the future if having the whole world loaded all the time proves to be
too demanding, but for not it's acceptable, and makes networking easier.